adapted the MultiselectComboBox for lazy data loading and introduced a JS connector
Adapted the MultiselectComboBox to support lazy loading. This was achieved following the implementation of the ComboBox component.
It invovled the introduction of the HasFilterableDataProvider interface, and the various supporting methods to achieve data lazy loading
and filtering. Furthermore, the latest version of the multiselect-combo-box web component was added in order to provide the pageSize property
which is crucial for supporting this feature.

The multiselect-combo-box-connector JS file was added as a frontend dependency. Similarly to before, this was achieved following the
implementation of the vaadinComboBoxConnector. This was done in such a way that the lazy loading logic was "delegated" to the
underlying vaadin-combo-box-light element that already extends the vaadin-data-provider-mixin.

As a last note, the unit tests were adapted to the new way of using the data provider and a new unit test for the page size property was added.