updated the demo view with a lazy loading example
Added one more demo to this view which showcases how lazy loading of data items is supported out of the box. The example generates a list of 1000 items which are set to the
MultiselectComboBox which has a default size of 50. This will ensure that the data is fetched page by page, thus demonstrating this feature.