feature/lazy loading (#16)
* adapted the MultiselectComboBox for lazy data loading and introduced a JS connector

Adapted the MultiselectComboBox to support lazy loading. This was achieved following the implementation of the ComboBox component.
It invovled the introduction of the HasFilterableDataProvider interface, and the various supporting methods to achieve data lazy loading
and filtering. Furthermore, the latest version of the multiselect-combo-box web component was added in order to provide the pageSize property
which is crucial for supporting this feature.

The multiselect-combo-box-connector JS file was added as a frontend dependency. Similarly to before, this was achieved following the
implementation of the vaadinComboBoxConnector. This was done in such a way that the lazy loading logic was "delegated" to the
underlying vaadin-combo-box-light element that already extends the vaadin-data-provider-mixin.

As a last note, the unit tests were adapted to the new way of using the data provider and a new unit test for the page size property was added.

* updated the demo view with a lazy loading example

Added one more demo to this view which showcases how lazy loading of data items is supported out of the box. The example generates a list of 1000 items which are set to the
MultiselectComboBox which has a default size of 50. This will ensure that the data is fetched page by page, thus demonstrating this feature.

* improved multiselect combo box api

Updated the class javadoc to better reflect its current capabilities.
Added more ways to construct this component, as well as additional
methods for setting items and data providers that better align with
the current ComboBox api.

* update multiselect combo box connectors

Introduced an ES6 and a legacy connector for the MultiselectComboBox,
in order to be compatible with V14 which still should work in
legacy browser projects.

* increased multiselect-combo-box dependency version

This version bump includes a critical update to the web component that
allows the focused index of the selected item to be preserved between
value selections. This prevents the overlay from randomly jumping
through the items between consecutive selections and provides an overall
improved usability.

* increased multiselect-combo-box dependency version to fix `focusedIndex`

Storing the focusedIndex on value changes had a side effect that after
clearing the items the focusedIndex is unchanged and the last
selected value becomes re-selected. To remedy this issue, the
multiselect-combo-box now resets the focusedIndex to -1 whenever the
value is cleared.

* increased multiselect-combo-box dependency version and project version

- Set the npm package dependency to the latest multiselect-combo-box web
component version 2.0.3
- Updated project version in pom.xml and readme to 2.1.0