introduced the `clear-button-visible` attribute
To better align with the new defaults for the vaadin components,
the `clear-button-visible` attribute is now also available to the
MultiselectComboBox.
With this update the component no longer displays the
clear icon by default, but instead it needs to be specified explicitly.
This change breaks the default behavior in favor of aligning better
with the current default behavior of the `vaadin-text-field` and
the `vaadin-combo-box` components.