feature/clear button visible attribute (#19)
* introduced the `clear-button-visible` attribute

To better align with the new defaults for the vaadin components,
the `clear-button-visible` attribute is now also available to the
MultiselectComboBox.
With this update the component no longer displays the
clear icon by default, but instead it needs to be specified explicitly.
This change breaks the default behavior in favor of aligning better
with the current default behavior of the `vaadin-text-field` and
the `vaadin-combo-box` components.

* exclude generated `flow-build-info.json` file from the add-on jar

Due to an issue in vaadin flow, the built jar for this component
contained the `flow-build-info.json` file which had local paths. This
further caused issues when using the MultiselectComboBox. o
To fix this, the exclusion for the `flow-build-info.json` is added
to the execution of the maven-jar-plugin.